Add files via upload
The goal of this project is to create regression models that are able to accurately estimate the price of the house given the features. In this project various Python tools are applied to get a visual understanding of the data and clean it to make it ready to apply machine learning models on it.